Move FVTTCompat / ModuleSettings to control.
