+++ src/control/index.js
@@ -1,5 +1,7 @@
 export * from './db/QuestDB.js';
 export * from './ui/index.js';
 export * from './FQLHooks.js';
+export * from './FVTTCompat.js';
+export * from './ModuleSettings.js';
 export * from './Socket.js';
 export * from './Utils.js';
